+++ test/test_unistd.rs
@@ -1118,3 +1118,23 @@ fn test_getpeereid_invalid_fd() {
     // getpeereid is not POSIX, so error codes are inconsistent between different Unices.
     assert!(getpeereid(-1).is_err());
 }
+
+#[cfg(any(target_os = "android", target_os = "freebsd", target_os = "linux", target_os = "openbsd"))]
+#[test]
+fn test_setresuid() {
+    setresuid(Some(getuid()), Some(getuid()), Some(getuid())).unwrap();
+    assert_eq!(geteuid(), getuid());
+
+    setresuid(None, None, None).unwrap();
+    assert_eq!(geteuid(), getuid());
+}
+
+#[cfg(any(target_os = "android", target_os = "freebsd",target_os = "linux", target_os = "openbsd"))]
+#[test]
+fn test_setresgid() {
+    setresgid(Some(getgid()), Some(getgid()), Some(getgid())).unwrap();
+    assert_eq!(getegid(), getgid());
+
+    setresgid(None, None, None).unwrap();
+    assert_eq!(getegid(), getgid());
+}